Reorganise the API
