+++ lib/src/api/engines/local/native.rs
@@ -1,4 +1,4 @@
-use crate::api::embedded::Db;
+use crate::api::engines::local::Db;
 use crate::api::opt::from_value;
 use crate::api::opt::DbResponse;
 use crate::api::opt::Param;
