+++ lib/src/api/method/mod.rs
@@ -191,11 +191,11 @@ where
 	/// use std::borrow::Cow;
 	/// use surrealdb::{Result, Surreal, StaticConnect};
 	/// use surrealdb::opt::auth::Root;
-	/// use surrealdb::protocol::Ws;
-	/// use surrealdb::net::WsClient;
+	/// use surrealdb::engines::remote::ws::Ws;
+	/// use surrealdb::engines::remote::ws::Client;
 	///
 	/// // Creates a new static instance of the client
-	/// static DB: Surreal<WsClient> = Surreal::new();
+	/// static DB: Surreal<Client> = Surreal::new();
 	///
 	/// #[derive(Serialize, Deserialize)]
 	/// struct Person {
@@ -232,7 +232,7 @@ where
 	/// use serde::{Serialize, Deserialize};
 	/// use std::borrow::Cow;
 	/// use surrealdb::Surreal;
-	/// use surrealdb::any::{Any, StaticConnect};
+	/// use surrealdb::engines::any::{Any, StaticConnect};
 	/// use surrealdb::opt::auth::Root;
 	///
 	/// // Creates a new static instance of the client
@@ -278,7 +278,7 @@ where
 	///
 	/// ```no_run
 	/// use surrealdb::Surreal;
-	/// use surrealdb::protocol::{Ws, Wss};
+	/// use surrealdb::engines::remote::{Ws, Wss};
 	///
 	/// # #[tokio::main]
 	/// # async fn main() -> surrealdb::Result<()> {
@@ -315,7 +315,7 @@ where
 	/// ```no_run
 	/// # #[tokio::main]
 	/// # async fn main() -> surrealdb::Result<()> {
-	/// # let db = surrealdb::any::connect("mem://").await?;
+	/// # let db = surrealdb::engines::any::connect("mem://").await?;
 	/// db.use_ns("test").use_db("test").await?;
 	/// # Ok(())
 	/// # }
@@ -342,7 +342,7 @@ where
 	///
 	/// # #[tokio::main]
 	/// # async fn main() -> surrealdb::Result<()> {
-	/// # let db = surrealdb::any::connect("mem://").await?;
+	/// # let db = surrealdb::engines::any::connect("mem://").await?;
 	/// #
 	/// // Assign the variable on the connection
 	/// db.set("name", Name {
@@ -382,7 +382,7 @@ where
 	///
 	/// # #[tokio::main]
 	/// # async fn main() -> surrealdb::Result<()> {
-	/// # let db = surrealdb::any::connect("mem://").await?;
+	/// # let db = surrealdb::engines::any::connect("mem://").await?;
 	/// #
 	/// // Assign the variable on the connection
 	/// db.set("name", Name {
@@ -427,7 +427,7 @@ where
 	///
 	/// # #[tokio::main]
 	/// # async fn main() -> surrealdb::Result<()> {
-	/// # let db = surrealdb::any::connect("mem://").await?;
+	/// # let db = surrealdb::engines::any::connect("mem://").await?;
 	/// #
 	/// // Sign in as root
 	/// db.signin(Root {
@@ -485,7 +485,7 @@ where
 	///
 	/// # #[tokio::main]
 	/// # async fn main() -> surrealdb::Result<()> {
-	/// # let db = surrealdb::any::connect("mem://").await?;
+	/// # let db = surrealdb::engines::any::connect("mem://").await?;
 	/// #
 	/// // Sign in as root
 	/// db.signin(Root {
@@ -520,7 +520,7 @@ where
 	///
 	/// # #[tokio::main]
 	/// # async fn main() -> surrealdb::Result<()> {
-	/// # let db = surrealdb::any::connect("mem://").await?;
+	/// # let db = surrealdb::engines::any::connect("mem://").await?;
 	/// #
 	/// // Sign in as root
 	/// db.signin(Root {
@@ -563,7 +563,7 @@ where
 	///
 	/// # #[tokio::main]
 	/// # async fn main() -> surrealdb::Result<()> {
-	/// # let db = surrealdb::any::connect("mem://").await?;
+	/// # let db = surrealdb::engines::any::connect("mem://").await?;
 	/// #
 	/// // Select the namespace/database to use
 	/// db.use_ns("namespace").use_db("database").await?;
@@ -601,7 +601,7 @@ where
 	/// ```no_run
 	/// # #[tokio::main]
 	/// # async fn main() -> surrealdb::Result<()> {
-	/// # let db = surrealdb::any::connect("mem://").await?;
+	/// # let db = surrealdb::engines::any::connect("mem://").await?;
 	/// db.invalidate().await?;
 	/// # Ok(())
 	/// # }
@@ -623,7 +623,7 @@ where
 	/// ```no_run
 	/// # #[tokio::main]
 	/// # async fn main() -> surrealdb::Result<()> {
-	/// # let db = surrealdb::any::connect("mem://").await?;
+	/// # let db = surrealdb::engines::any::connect("mem://").await?;
 	/// # let token = String::new();
 	/// db.authenticate(token).await?;
 	/// # Ok(())
@@ -645,7 +645,7 @@ where
 	/// # struct Person;
 	/// # #[tokio::main]
 	/// # async fn main() -> surrealdb::Result<()> {
-	/// # let db = surrealdb::any::connect("mem://").await?;
+	/// # let db = surrealdb::engines::any::connect("mem://").await?;
 	/// #
 	/// // Run queries
 	/// let mut result = db
@@ -681,7 +681,7 @@ where
 	/// #
 	/// # #[tokio::main]
 	/// # async fn main() -> surrealdb::Result<()> {
-	/// # let db = surrealdb::any::connect("mem://").await?;
+	/// # let db = surrealdb::engines::any::connect("mem://").await?;
 	/// #
 	/// // Select all records from a table
 	/// let people: Vec<Person> = db.select("person").await?;
@@ -728,7 +728,7 @@ where
 	///
 	/// # #[tokio::main]
 	/// # async fn main() -> surrealdb::Result<()> {
-	/// # let db = surrealdb::any::connect("mem://").await?;
+	/// # let db = surrealdb::engines::any::connect("mem://").await?;
 	/// #
 	/// // Create a record with a random ID
 	/// let person: Person = db.create("person").await?;
@@ -781,7 +781,7 @@ where
 	///
 	/// # #[tokio::main]
 	/// # async fn main() -> surrealdb::Result<()> {
-	/// # let db = surrealdb::any::connect("mem://").await?;
+	/// # let db = surrealdb::engines::any::connect("mem://").await?;
 	/// #
 	/// // Update all records in a table
 	/// let people: Vec<Person> = db.update("person").await?;
@@ -828,7 +828,7 @@ where
 	///
 	/// # #[tokio::main]
 	/// # async fn main() -> surrealdb::Result<()> {
-	/// # let db = surrealdb::any::connect("mem://").await?;
+	/// # let db = surrealdb::engines::any::connect("mem://").await?;
 	/// #
 	/// // Update all records in a table
 	/// let people: Vec<Person> = db.update("person")
@@ -879,7 +879,7 @@ where
 	///
 	/// # #[tokio::main]
 	/// # async fn main() -> surrealdb::Result<()> {
-	/// # let db = surrealdb::any::connect("mem://").await?;
+	/// # let db = surrealdb::engines::any::connect("mem://").await?;
 	/// #
 	/// // Update all records in a table
 	/// let people: Vec<Person> = db.update("person")
@@ -912,7 +912,7 @@ where
 	/// ```no_run
 	/// # #[tokio::main]
 	/// # async fn main() -> surrealdb::Result<()> {
-	/// # let db = surrealdb::any::connect("mem://").await?;
+	/// # let db = surrealdb::engines::any::connect("mem://").await?;
 	/// #
 	/// // Delete all records from a table
 	/// db.delete("person").await?;
@@ -939,7 +939,7 @@ where
 	/// ```no_run
 	/// # #[tokio::main]
 	/// # async fn main() -> surrealdb::Result<()> {
-	/// # let db = surrealdb::any::connect("mem://").await?;
+	/// # let db = surrealdb::engines::any::connect("mem://").await?;
 	/// let version = db.version().await?;
 	/// # Ok(())
 	/// # }
@@ -957,7 +957,7 @@ where
 	/// ```no_run
 	/// # #[tokio::main]
 	/// # async fn main() -> surrealdb::Result<()> {
-	/// # let db = surrealdb::any::connect("mem://").await?;
+	/// # let db = surrealdb::engines::any::connect("mem://").await?;
 	/// db.health().await?;
 	/// # Ok(())
 	/// # }
@@ -995,7 +995,7 @@ where
 	/// ```no_run
 	/// # #[tokio::main]
 	/// # async fn main() -> surrealdb::Result<()> {
-	/// # let db = surrealdb::any::connect("mem://").await?;
+	/// # let db = surrealdb::engines::any::connect("mem://").await?;
 	/// db.export("backup.sql").await?;
 	/// # Ok(())
 	/// # }
@@ -1021,7 +1021,7 @@ where
 	/// ```no_run
 	/// # #[tokio::main]
 	/// # async fn main() -> surrealdb::Result<()> {
-	/// # let db = surrealdb::any::connect("mem://").await?;
+	/// # let db = surrealdb::engines::any::connect("mem://").await?;
 	/// db.import("backup.sql").await?;
 	/// # Ok(())
 	/// # }
