+++ lib/src/api/method/tests/mod.rs
@@ -6,12 +6,12 @@ mod server;
 mod types;
 
 use crate::api::method::tests::types::AuthParams;
+use crate::api::opt::auth::Database;
 use crate::api::opt::auth::Jwt;
+use crate::api::opt::auth::NameSpace;
 use crate::api::opt::auth::Root;
-use crate::api::opt::Database;
-use crate::api::opt::NameSpace;
+use crate::api::opt::auth::Scope;
 use crate::api::opt::PatchOp;
-use crate::api::opt::Scope;
 use crate::api::QueryResponse;
 use crate::api::StaticConnect;
 use crate::api::Surreal;
