+++ lib/src/api/mod.rs
@@ -1,49 +1,6 @@
-pub mod any;
+pub mod engines;
 pub mod method;
-
-#[cfg(any(
-	feature = "kv-mem",
-	feature = "kv-tikv",
-	feature = "kv-rocksdb",
-	feature = "kv-fdb",
-	feature = "kv-indxdb",
-))]
-#[cfg_attr(
-	docsrs,
-	doc(cfg(any(
-		feature = "kv-mem",
-		feature = "kv-tikv",
-		feature = "kv-rocksdb",
-		feature = "kv-fdb",
-		feature = "kv-indxdb",
-	)))
-)]
-pub mod embedded;
-#[cfg(any(feature = "protocol-http", feature = "protocol-ws"))]
-#[cfg_attr(docsrs, doc(cfg(any(feature = "protocol-http", feature = "protocol-ws"))))]
-pub mod net;
 pub mod opt;
-#[cfg(any(feature = "protocol-http", feature = "protocol-ws"))]
-#[cfg_attr(docsrs, doc(cfg(any(feature = "protocol-http", feature = "protocol-ws"))))]
-pub mod protocol;
-#[cfg(any(
-	feature = "kv-mem",
-	feature = "kv-tikv",
-	feature = "kv-rocksdb",
-	feature = "kv-fdb",
-	feature = "kv-indxdb",
-))]
-#[cfg_attr(
-	docsrs,
-	doc(cfg(any(
-		feature = "kv-mem",
-		feature = "kv-tikv",
-		feature = "kv-rocksdb",
-		feature = "kv-fdb",
-		feature = "kv-indxdb",
-	)))
-)]
-pub mod storage;
 
 pub(super) mod err;
 
@@ -54,17 +11,6 @@ use crate::api::opt::DbResponse;
 use crate::api::opt::Param;
 use crate::api::opt::ServerAddrs;
 use crate::api::opt::ToServerAddrs;
-use crate::sql::statements::CreateStatement;
-use crate::sql::statements::DeleteStatement;
-use crate::sql::statements::SelectStatement;
-use crate::sql::statements::UpdateStatement;
-use crate::sql::Array;
-use crate::sql::Data;
-use crate::sql::Field;
-use crate::sql::Fields;
-use crate::sql::Output;
-use crate::sql::Value;
-use crate::sql::Values;
 use flume::Receiver;
 use flume::Sender;
 use method::Method;
@@ -77,7 +23,6 @@ use std::fmt::Debug;
 use std::future::Future;
 use std::future::IntoFuture;
 use std::marker::PhantomData;
-use std::mem;
 use std::pin::Pin;
 use std::sync::atomic::AtomicI64;
 use std::sync::atomic::Ordering;
@@ -181,7 +126,7 @@ where
 	/// ```no_run
 	/// # #[tokio::main]
 	/// # async fn main() -> surrealdb::Result<()> {
-	/// use surrealdb::protocol::Ws;
+	/// use surrealdb::engines::remote::ws::Ws;
 	/// use surrealdb::Surreal;
 	///
 	/// let db = Surreal::connect::<Ws>("localhost:8000")
@@ -349,112 +294,5 @@ where
 	fn extract(&self) -> Result<&Router<C>> {
 		let router = self.get().ok_or(Error::ConnectionUninitialised)?;
 		Ok(router)
-	}
-}
-
-#[allow(dead_code)] // used by the the embedded database and `http`
-fn split_params(params: &mut [Value]) -> (bool, Values, Value) {
-	let (what, data) = match params {
-		[what] => (mem::take(what), Value::None),
-		[what, data] => (mem::take(what), mem::take(data)),
-		_ => unreachable!(),
-	};
-	let one = what.is_thing();
-	let what = match what {
-		Value::Array(Array(vec)) => Values(vec),
-		value => Values(vec![value]),
-	};
-	(one, what, data)
-}
-
-#[allow(dead_code)] // used by the the embedded database and `http`
-fn create_statement(params: &mut [Value]) -> CreateStatement {
-	let (_, what, data) = split_params(params);
-	let data = match data {
-		Value::None | Value::Null => None,
-		value => Some(Data::ContentExpression(value)),
-	};
-	CreateStatement {
-		what,
-		data,
-		output: Some(Output::After),
-		..Default::default()
-	}
-}
-
-#[allow(dead_code)] // used by the the embedded database and `http`
-fn update_statement(params: &mut [Value]) -> (bool, UpdateStatement) {
-	let (one, what, data) = split_params(params);
-	let data = match data {
-		Value::None | Value::Null => None,
-		value => Some(Data::ContentExpression(value)),
-	};
-	(
-		one,
-		UpdateStatement {
-			what,
-			data,
-			output: Some(Output::After),
-			..Default::default()
-		},
-	)
-}
-
-#[allow(dead_code)] // used by the the embedded database and `http`
-fn patch_statement(params: &mut [Value]) -> (bool, UpdateStatement) {
-	let (one, what, data) = split_params(params);
-	let data = match data {
-		Value::None | Value::Null => None,
-		value => Some(Data::PatchExpression(value)),
-	};
-	(
-		one,
-		UpdateStatement {
-			what,
-			data,
-			output: Some(Output::Diff),
-			..Default::default()
-		},
-	)
-}
-
-#[allow(dead_code)] // used by the the embedded database and `http`
-fn merge_statement(params: &mut [Value]) -> (bool, UpdateStatement) {
-	let (one, what, data) = split_params(params);
-	let data = match data {
-		Value::None | Value::Null => None,
-		value => Some(Data::MergeExpression(value)),
-	};
-	(
-		one,
-		UpdateStatement {
-			what,
-			data,
-			output: Some(Output::After),
-			..Default::default()
-		},
-	)
-}
-
-#[allow(dead_code)] // used by the the embedded database and `http`
-fn select_statement(params: &mut [Value]) -> (bool, SelectStatement) {
-	let (one, what, _) = split_params(params);
-	(
-		one,
-		SelectStatement {
-			what,
-			expr: Fields(vec![Field::All]),
-			..Default::default()
-		},
-	)
-}
-
-#[allow(dead_code)] // used by the the embedded database and `http`
-fn delete_statement(params: &mut [Value]) -> DeleteStatement {
-	let (_, what, _) = split_params(params);
-	DeleteStatement {
-		what,
-		output: Some(Output::None),
-		..Default::default()
 	}
 }
