+++ lib/src/api/opt/server_addrs/rocksdb.rs
@@ -1,10 +1,10 @@
-use crate::api::embedded::Db;
+use crate::api::engines::local::Db;
+use crate::api::engines::local::File;
+use crate::api::engines::local::RocksDb;
 use crate::api::err::Error;
 use crate::api::opt::ServerAddrs;
 use crate::api::opt::Strict;
 use crate::api::opt::ToServerAddrs;
-use crate::api::storage::File;
-use crate::api::storage::RocksDb;
 use crate::api::Result;
 use std::path::Path;
 use url::Url;
